+++ python/src/distance/kohonen.cpp
