+++ python/src/mapping/KOC.cpp
@@ -1,6 +1,5 @@
 #include "metric_py.hpp"
 #include "modules/mapping.hpp"
-#include "modules/utils/graph/graph.cpp"
 
 #include <boost/python.hpp>
