+++ python/src/mapping/affprop.cpp
