+++ python/src/mapping/autoencoder.cpp
