--- python/src/metric_VOI.hpp
@@ -1,11 +1,0 @@
-/*
-  This Source Code Form is subject to the terms of the Mozilla Public
-  License, v. 2.0. If a copy of the MPL was not distributed with this
-  file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-  Copyright (c) 2020 Panda Team
-*/
-
-#pragma once
-namespace metric_VOI {
-};
