+++ python/src/metric_py.hpp
@@ -85,5 +85,14 @@ public:
 };
 
 
-std::string getObjType(const boost::python::api::object& obj);
-std::string getObjType(PyObject* obj_ptr);
+inline std::string getObjType(const boost::python::api::object& obj) {
+    boost::python::extract<boost::python::object> objectExtractor(obj);
+    boost::python::object o=objectExtractor();
+    std::string obj_type = boost::python::extract<std::string>(o.attr("__class__").attr("__name__"));
+    return obj_type;
+}
+
+inline std::string getObjType(PyObject* obj_ptr) {
+       boost::python::object obj(boost::python::handle<>(boost::python::borrowed(obj_ptr)));
+       return getObjType(obj);
+}
