+++ python/src/space/tree.cpp
