+++ python/src/todo/SSIM.cpp
